samples/chat-gpt/pyproject.toml
[project]
name = "chatgpt-fastapi"
version = "1.0.0"
description = "A FastAPI application with OpenAI ChatGPT integration"
authors = [
    {name = "ChatGPT FastAPI", email = "example@example.com"}
]
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "openai==1.54.4",
    "pydantic==2.10.4",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.12",
    "surrealdb==0.3.2",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []